Safe mode and power-on sequencing
In order to stay on the board must set HOLD(GPIO4) high early during
boot.

When running off battery the power-on button must be held for 2s to
power things on, but if it's held for too long the board powers off. We
need to disable the 1s safe mode timer in order to give the user a long
enough window to set the pins up so the button can be released and the
board stay powered on. Safe mode can be entered by holding BTN_A during
boot.