Collect doc files based primarily on an inclusion list
It frequently arises that files were inadvertently included in the doc
build, because it was based on a series of exclude patterns, but new
filenames that don't match the exclude patterns are frequently created
by new build systems, new submodules, un-tidy developer source trees, etc.

So, use include_patterns as a set of positive patterns, then weed out
a few specific items that otherwise match the include patterns.