espressif: don't enable AUDIOMP3 without AUDIOCORE
there are some lines above that set it OFF if needed; the top level
logic will set it ON if audiocore is enabled.

Before this, some boards without any audio capability (e.g., pycamera
with esp32s3, lacking both audioio and i2s) enabled mp3. This was not
useful, and with other changes it will become a build error.