ports: Fix m_realloc/m_free when MICROPY_MALLOC_USES_ALLOCATED_SIZE is set
When 'MICROPY_MALLOC_USES_ALLOCATED_SIZE' is set, we need to pass the
previously allocated size to 'm_realloc', 'm_realloc_maybe' and 'm_free'.

For the atmel-samd and raspberrypi ports, the variable keeping track of the
allocated buffer size was already present, but not properly used.
For the nordic and stm ports, there was no such veriable. It is now added
only if necessary to save on resources.

Signed-off-by: Samantaz Fox <coding@samantaz.fr>